server/requirements.txt
# FastAPI & Web
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.2

# LiveKit
livekit-agents[groq,silero,turn-detector]~=1.3

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.2
langchain-text-splitters>=0.3

# Firestore
firebase-admin==6.6.0
livekit-plugins-deepgram
